NO-ISSUE: test: merge image-ops.ts into container.ts
image-ops.ts duplicated push functionality already in container.ts
(both push single images, both use skopeo for multi-arch). Move the
only unique function (orasAttach) into container.ts and delete the
redundant file.
